+++ test/core-download/downloadFile.test.js
@@ -1,0 +1,16 @@
+const cd0 = require('../../lib/core-download')
+
+test('download/downloadFile', () => {
+  const fileWorkingObject = {
+    client: {
+      downloadFile: jest.fn().mockResolvedValue()
+    },
+    localFile: './test.txt',
+    remoteFile: './test.txt'
+  }
+
+  cd0.downloadFile(fileWorkingObject)
+    .then(() => {
+      expect(fileWorkingObject.client.downloadFile).toBeCalled()
+    })
+})